requirements.txt
torch == 2.2.2
numpy == 1.26.4
openpyxl == 3.1.5
pandas == 2.2.1
xlrd == 2.0.2
openpyxl == 3.1.5
scikit-learn == 1.6.1
tqdm

timm
hyperimpute
tabulate

# tab-ddpm requirements
catboost==1.0.3
category-encoders==2.3.0
dython==0.5.1
icecream==2.1.2
libzero==0.0.8
numpy==1.21.4
optuna==2.10.1
pandas==1.3.4
pyarrow==6.0.0
rtdl==0.0.9
scikit-learn==1.0.2
scipy==1.7.2
skorch==0.11.0
tomli-w==0.4.0
tomli==1.2.2
tqdm==4.62.3

# smote
imbalanced-learn==0.7.0

# tvae
rdt==0.6.4 